story(ccls-2026): split integration tests
