+++ caab-service/src/integrationTest/java/uk/gov/laa/ccms/data/api/controller/BaseApplicationControllerIntegrationTest.java
@@ -4,36 +4,22 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static uk.gov.laa.ccms.caab.api.audit.AuditorAwareImpl.currentUserHolder;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.net.URI;
 import java.util.List;
-import java.util.Objects;
-import java.util.Random;
 import java.util.stream.Stream;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.jdbc.Sql;
 import uk.gov.laa.ccms.caab.api.controller.ApplicationController;
-import uk.gov.laa.ccms.caab.api.controller.LinkedCaseController;
-import uk.gov.laa.ccms.caab.api.controller.OpponentController;
-import uk.gov.laa.ccms.caab.api.controller.PriorAuthorityController;
-import uk.gov.laa.ccms.caab.api.controller.ProceedingController;
-import uk.gov.laa.ccms.caab.api.controller.ScopeLimitationController;
-import uk.gov.laa.ccms.caab.api.entity.AuditTrail;
 import uk.gov.laa.ccms.caab.api.service.ApplicationService;
 import uk.gov.laa.ccms.caab.model.ApplicationDetail;
 import uk.gov.laa.ccms.caab.model.ApplicationDetails;
@@ -45,62 +31,15 @@ import uk.gov.laa.ccms.caab.model.PriorAuthority;
 import uk.gov.laa.ccms.caab.model.Proceeding;
 import uk.gov.laa.ccms.caab.model.ScopeLimitation;
 
-public abstract class BaseApplicationControllerIntegrationTest {
+public abstract class BaseApplicationControllerIntegrationTest
+    extends AbstractControllerIntegrationTest {
 
   @Autowired
   private ApplicationController applicationController;
 
   @Autowired
-  private LinkedCaseController linkedCaseController;
-
-  @Autowired
-  private OpponentController opponentController;
-
-  @Autowired
-  private PriorAuthorityController priorAuthorityController;
-
-  @Autowired
-  private ProceedingController proceedingController;
-
-  @Autowired
-  private ScopeLimitationController scopeLimitationController;
-
-  @Autowired
   private ApplicationService applicationService;
 
-  private final String caabUserLoginId = "audit@user.com";
-
-  @BeforeEach
-  public void setup() {
-    currentUserHolder.set(caabUserLoginId);
-  }
-
-  /**
-   * Loads the JSON file from the classpath and parses it into a specified object.
-   *
-   * @param jsonFilePath The path to the JSON file to load
-   */
-  private <T> T loadObjectFromJson(String jsonFilePath, Class<T> objectType) throws IOException {
-    ObjectMapper objectMapper = new ObjectMapper();
-    ClassPathResource resource = new ClassPathResource(jsonFilePath);
-    return objectMapper.readValue(resource.getInputStream(), objectType);
-  }
-
-
-  /**
-   * Generates a random 12-digit number string to be used as a case reference number.
-   */
-  private String generateTestCaseRef(){
-    Random random = new Random();
-    StringBuilder stringBuilder = new StringBuilder(12);
-
-    for (int i = 0; i < 12; i++) {
-      int digit = random.nextInt(10); // Generate a random digit (0-9)
-      stringBuilder.append(digit);
-    }
-
-    return stringBuilder.toString();
-  }
 
   private static Stream<Arguments> createApplicationUpdateArguments() {
     return Stream.of(
@@ -151,40 +90,6 @@ public abstract class BaseApplicationControllerIntegrationTest {
         ApplicationDetail.class, List.of(fieldToIgnore, "auditTrail")));
   }
 
-  // Utility method to resolve the method chain and get the value
-  private Object getObjectToCheckFromMethod(Object initialObject, String[] methodCalls)
-      throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
-
-    Object currentObject = initialObject;
-    for (String methodName : methodCalls) {
-      Method method = currentObject.getClass().getMethod(methodName);
-      currentObject = method.invoke(currentObject);
-    }
-    return currentObject;
-  }
-
-
-  public static boolean areAllFieldsEqual(Object obj1, Object obj2, Class<?> clazz, List<String> fieldsToIgnore)
-      throws IllegalArgumentException, IllegalAccessException {
-
-    for (Field field : clazz.getDeclaredFields()) {
-      field.setAccessible(true);
-
-      // Continue to the next field if it should be ignored
-      if (fieldsToIgnore.contains(field.getName())) {
-        continue;
-      }
-
-      Object value1 = field.get(obj1);
-      Object value2 = field.get(obj2);
-
-      if (!Objects.equals(value1, value2)) {
-        System.out.println("Field " + field.getName() + " is not equal: " + value1 + " != " + value2);
-        return false;
-      }
-    }
-    return true;
-  }
 
   /**
    * Arguments for the testCreateApplication test method.
@@ -488,18 +393,6 @@ public abstract class BaseApplicationControllerIntegrationTest {
 
   @Test
   @Sql(scripts = {"/sql/application_insert.sql", "/sql/proceeding_insert.sql"})
-  public void updateProceeding() throws IOException {
-    Long proceedingId = 2L;
-
-    Proceeding updatedProceeding = loadObjectFromJson("/json/proceeding_new.json", Proceeding.class);
-
-    ResponseEntity<Void> response = proceedingController.updateProceeding(proceedingId, caabUserLoginId, updatedProceeding);
-
-    assertEquals(HttpStatus.NO_CONTENT, response.getStatusCode());
-  }
-
-  @Test
-  @Sql(scripts = {"/sql/application_insert.sql", "/sql/proceeding_insert.sql"})
   public void getApplicationProceedings() {
     Long caseRef = 41L;
 
@@ -510,22 +403,6 @@ public abstract class BaseApplicationControllerIntegrationTest {
 
     List<Proceeding> proceedings = responseEntity.getBody();
     assertEquals(1, proceedings.size());
-  }
-
-  @Test
-  @Sql(scripts = {"/sql/application_insert.sql", "/sql/proceeding_insert.sql"})
-  public void deleteApplicationProceedings() {
-    Long caseRef = 41L;
-    Long proceedingRef = 2L;
-
-    proceedingController.removeProceeding(proceedingRef, caabUserLoginId);
-    ResponseEntity<List<Proceeding>> responseEntity = applicationController.getApplicationProceedings(caseRef);
-
-    assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
-    assertNotNull(responseEntity.getBody());
-
-    List<Proceeding> proceedings = responseEntity.getBody();
-    assertEquals(0, proceedings.size());
   }
 
   @Test
@@ -566,22 +443,6 @@ public abstract class BaseApplicationControllerIntegrationTest {
   }
 
   @Test
-  @Sql(scripts = {"/sql/application_insert.sql", "/sql/linked_cases_insert.sql"})
-  public void deleteLinkedCase() {
-    Long caseRef = 41L;
-    Long linkedCaseRef = 2L;
-
-    linkedCaseController.removeLinkedCase(linkedCaseRef, caabUserLoginId);
-    ResponseEntity<List<LinkedCase>> responseEntity = applicationController.getApplicationLinkedCases(caseRef);
-
-    assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
-    assertNotNull(responseEntity.getBody());
-
-    List<LinkedCase> linkedCases = responseEntity.getBody();
-    assertEquals(0, linkedCases.size());
-  }
-
-  @Test
   @Sql(scripts = {"/sql/application_insert.sql"})
   public void addPriorAuthority() throws IOException {
     Long caseRef = 41L;
@@ -596,18 +457,6 @@ public abstract class BaseApplicationControllerIntegrationTest {
 
   @Test
   @Sql(scripts = {"/sql/application_insert.sql", "/sql/prior_authority_insert.sql"})
-  public void updatePriorAuthority() throws IOException {
-    Long priorAuthorityId = 2L;
-
-    PriorAuthority updatedPriorAuthority = loadObjectFromJson("/json/prior_authority_new.json", PriorAuthority.class);
-
-    ResponseEntity<Void> response = priorAuthorityController.updatePriorAuthority(priorAuthorityId, caabUserLoginId, updatedPriorAuthority);
-
-    assertEquals(HttpStatus.NO_CONTENT, response.getStatusCode());
-  }
-
-  @Test
-  @Sql(scripts = {"/sql/application_insert.sql", "/sql/prior_authority_insert.sql"})
   public void getPriorAuthority() {
     Long caseRef = 41L;
 
@@ -618,84 +467,6 @@ public abstract class BaseApplicationControllerIntegrationTest {
 
     List<PriorAuthority> priorAuthorities = responseEntity.getBody();
     assertEquals(1, priorAuthorities.size());
-  }
-
-  @Test
-  @Sql(scripts = {"/sql/application_insert.sql", "/sql/prior_authority_insert.sql"})
-  public void deletePriorAuthority() {
-    Long caseRef = 41L;
-    Long priorAuthorityRef = 2L;
-
-    priorAuthorityController.removePriorAuthority(priorAuthorityRef, caabUserLoginId);
-    ResponseEntity<List<PriorAuthority>> responseEntity = applicationController.getApplicationPriorAuthorities(caseRef);
-
-    assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
-    assertNotNull(responseEntity.getBody());
-
-    List<PriorAuthority> priorAuthorities = responseEntity.getBody();
-    assertEquals(0, priorAuthorities.size());
-  }
-
-  @Test
-  @Sql(scripts = {"/sql/application_insert.sql", "/sql/proceeding_insert.sql"})
-  public void addScopeLimitationToProceeding() throws IOException {
-    Long proceedingId = 2L;
-
-    ScopeLimitation scopeLimitation = loadObjectFromJson("/json/scope_limitation_new.json", ScopeLimitation.class);
-
-    ResponseEntity<Void> response = proceedingController.addProceedingScopeLimitation(proceedingId, caabUserLoginId, scopeLimitation);
-
-    assertEquals(HttpStatus.CREATED, response.getStatusCode());
-  }
-
-  @Test
-  @Sql(scripts = {"/sql/application_insert.sql", "/sql/proceeding_insert.sql", "/sql/scope_limitation_insert.sql"})
-  public void updateScopeLimitationInProceeding() throws IOException {
-    Long scopeLimitationId = 3L;
-
-    ScopeLimitation updatedScopeLimitation = loadObjectFromJson("/json/scope_limitation_new.json", ScopeLimitation.class);
-
-    ResponseEntity<Void> response = scopeLimitationController.updateScopeLimitation(scopeLimitationId, caabUserLoginId, updatedScopeLimitation);
-
-    assertEquals(HttpStatus.NO_CONTENT, response.getStatusCode());
-  }
-
-
-  @Test
-  @Sql(scripts = {
-      "/sql/application_insert.sql",
-      "/sql/proceeding_insert.sql",
-      "/sql/scope_limitation_insert.sql"})
-  public void getScopeLimitationsForProceeding() {
-    Long proceedingId = 2L;
-
-    ResponseEntity<List<ScopeLimitation>> responseEntity = proceedingController.getProceedingsScopeLimitations(proceedingId);
-
-    assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
-    assertNotNull(responseEntity.getBody());
-
-    List<ScopeLimitation> scopeLimitations = responseEntity.getBody();
-    assertFalse(scopeLimitations.isEmpty());
-  }
-
-  @Test
-  @Sql(scripts = {
-      "/sql/application_insert.sql",
-      "/sql/proceeding_insert.sql",
-      "/sql/scope_limitation_insert.sql"})
-  public void removeScopeLimitationFromProceeding() {
-    Long scopeLimitationId = 3L;
-
-    ResponseEntity<Void> response = scopeLimitationController.removeScopeLimitation(scopeLimitationId, caabUserLoginId);
-    assertEquals(HttpStatus.NO_CONTENT, response.getStatusCode());
-
-    ResponseEntity<List<ScopeLimitation>> responseEntity = proceedingController.getProceedingsScopeLimitations(2L);
-
-    assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
-    assertNotNull(responseEntity.getBody());
-
-    List<ScopeLimitation> scopeLimitations = responseEntity.getBody();
-    assertTrue(scopeLimitations.isEmpty());
   }
 
   @Test
@@ -736,19 +507,6 @@ public abstract class BaseApplicationControllerIntegrationTest {
   }
 
   @Test
-  @Sql(scripts = {"/sql/application_insert.sql", "/sql/opponent_insert.sql"})
-  public void updateOpponentInApplication() throws IOException {
-    Long opponentId = 3L;
-
-    Opponent updatedOpponent = loadObjectFromJson("/json/opponent_new.json", Opponent.class);
-
-    ResponseEntity<Void> response = opponentController.updateOpponent(opponentId, caabUserLoginId, updatedOpponent);
-
-    assertEquals(HttpStatus.NO_CONTENT, response.getStatusCode());
-  }
-
-
-  @Test
   @Sql(scripts = {
       "/sql/application_insert.sql",
       "/sql/opponent_insert.sql"})
@@ -762,123 +520,6 @@ public abstract class BaseApplicationControllerIntegrationTest {
 
     List<Opponent> opponents = responseEntity.getBody();
     assertFalse(opponents.isEmpty());
-  }
-
-  @Test
-  @Sql(scripts = {
-      "/sql/application_insert.sql",
-      "/sql/opponent_insert.sql"})
-  public void removeOpponentFromApplication() {
-    Long caseRef = 41L;
-    Long opponentRef = 3L;
-
-    ResponseEntity<Void> response = opponentController.removeOpponent(
-        opponentRef, caabUserLoginId);
-    assertEquals(HttpStatus.NO_CONTENT, response.getStatusCode());
-
-    ResponseEntity<List<Opponent>> responseEntity = applicationController.getApplicationOpponents(
-        caseRef);
-
-    assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
-    assertNotNull(responseEntity.getBody());
-
-    List<Opponent> opponents = responseEntity.getBody();
-    assertTrue(opponents.isEmpty());
-  }
-
-  /**
-   * Assert that the audit trail is set correctly in the object. Uses reflection to navigate the
-   * auditTrailMethod path and assert that the audit trail is set correctly.
-   */
-  private void assertAuditTrail(Object object, String auditMethod, String expectedAuditUser)
-      throws InvocationTargetException, IllegalAccessException, NoSuchMethodException {
-    String[] methodCalls = auditMethod.split("\\.");
-    assertAuditTrailRecursive(object, methodCalls, 0, expectedAuditUser);
-  }
-
-  /**
-   * Recursively navigate the object using the methodCalls array and assert that the audit trail is
-   * set correctly.
-   */
-  private void assertAuditTrailRecursive(Object object, String[] methodCalls, int index, String expectedAuditUser)
-      throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
-    if (index >= methodCalls.length || object == null) {
-      return;
-    }
-
-    Method method = object.getClass().getMethod(methodCalls[index]);
-    object = method.invoke(object);
-
-    if (object instanceof List<?> list) {
-      for (Object item : list) {
-        assertAuditTrailRecursive(item, methodCalls, index + 1, expectedAuditUser);
-      }
-    } else {
-      if (index == methodCalls.length - 1) {
-        assertAuditDetailsInObject(object, expectedAuditUser);
-      } else {
-        assertAuditTrailRecursive(object, methodCalls, index + 1, expectedAuditUser);
-      }
-    }
-  }
-
-  /**
-   * Used to assert that the audit trail is set correctly in the object.
-   *
-   * @param object The object to check
-   * @param expectedAuditUser The expected audit user
-   */
-  private void assertAuditDetailsInObject(Object object, String expectedAuditUser) {
-    if (object instanceof AuditTrail auditTrail) {
-      assertNotNull(auditTrail);
-      assertEquals(expectedAuditUser, auditTrail.getCreatedBy());
-      assertEquals(expectedAuditUser, auditTrail.getLastSavedBy());
-      assertNotNull(auditTrail.getCreated());
-      assertNotNull(auditTrail.getLastSaved());
-    }
-  }
-
-  /**
-   * Sets the audit trail properties to null in the object. Uses reflection to navigate the
-   * list of auditTrailsToNull and set the audit trail to null.
-   */
-  private void setAuditPropertiesToNull(Object object, List<String> auditTrailsToNull) throws NoSuchFieldException, IllegalAccessException {
-    if (object == null || auditTrailsToNull == null) {
-      return;
-    }
-    for (String trail : auditTrailsToNull) {
-      String[] fields = trail.split("\\.");
-      nullifyFieldRecursive(object, fields, 0);
-    }
-  }
-
-  /**
-   * Recursively navigate the object using the fields array and set the field to null.
-   */
-  private static void nullifyFieldRecursive(Object object, String[] fields, int index)
-      throws NoSuchFieldException, IllegalAccessException {
-    if (object == null || index >= fields.length) {
-      return;
-    }
-
-    if (object instanceof List<?> list) {
-      // Handle List objects
-      for (Object item : list) {
-        nullifyFieldRecursive(item, fields, index);
-      }
-    } else {
-      Field field = object.getClass().getDeclaredField(fields[index]);
-      field.setAccessible(true);
-
-      if (index == fields.length - 1) {
-        // Last field in the path, set to null
-        field.set(object, null);
-      } else {
-        // Not the last field, navigate to the next object or list in the path
-        Object nextObject = field.get(object);
-        nullifyFieldRecursive(nextObject, fields, index + 1);
-      }
-    }
   }
 
 }
